Add typescript type definitions
This file is necessary for typescript projects to be able to use
react-keyed-file-browser without writing their own definition files.
This is based on the work of @graywolf336 and adds definitions for the
FileBrowser and its props.

The definition file doesn't provide type definitions for rendererers
implementations but they should be the same as their interface.

Closes #44.